requirements.txt
google-api-core==1.31.1
google-api-python-client==1.7.11
google-auth==1.34.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.5
google-cloud-bigquery==2.24.0
google-cloud-bigquery-storage==2.6.3
google-cloud-core==1.7.2
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
googleauthentication==0.0.15
pandas==1.3.1
python-dotenv==0.19.0
numpy==1.21.2